requirements.txt
django
psycopg2-binary
pillow
redis
easyocr
opencv-python-headless
requests  # For the NHTSA API
torch
torchvision